optimized memory allocation (#8)

+++ src/io/io.h
@@ -22,7 +22,7 @@ namespace arboretum {
       thrust::host_vector<float, thrust::cuda::experimental::pinned_allocator< float > > grad;
       size_t rows;
       size_t columns;
-      void Init(const float initial_y, std::function<double(const double, const double)> func);
+      void Init(const float initial_y, std::function<float(const float, const float)> func);
       void UpdateGrad();
       DataMatrix(int rows, int columns);
     private:
